RL-plugings merged + revised EWCRL plugin
